% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ENCODExplorer.R
\docType{data}
\name{accession_df}
\alias{accession_df}
\title{Metadata from the ENCODE database in a R object.
The tables were generated with the \code{prepare_ENCODEdb} function.}
\format{data frame}
\usage{
data(accession_df)
}
\value{
data frame
}
\description{
Metadata from the ENCODE database in a R object.
The tables were generated with the \code{prepare_ENCODEdb} function.
}
\seealso{
\code{\link{get_encode_types}} to get a list of possible types. Note
  that some of the types are empty tables that are not included in the
  database created with \code{\link{prepare_ENCODEdb}} function.
}
\keyword{datasets}

