% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/summaryBuilders.R
\name{buildConsensusPeaks}
\alias{buildConsensusPeaks}
\title{Calculates the consensus peaks defined by the results of a previously
completed ENCODE query.}
\usage{
buildConsensusPeaks(query_results, split_by = NULL,
  consensus_threshold = 1, simplify = FALSE, temp_dir = ".",
  force = FALSE)
}
\arguments{
\item{query_results}{A data.table returned by \code{\link{queryEncode}} or 
\code{\link{queryEncodeGeneric}}.}

\item{split_by}{A vector of column names from query_results that will be used
to split the consensus binding sites according to condition.
If \code{NULL}, all elements of query_results are used in the
same consensus calculation.}

\item{consensus_threshold}{A numeric value between 0 and 1, indicating the
proportion of peak files in which a peak must
appear for it to be included within the consensus.}

\item{simplify}{If TRUE, non-discriminatory columns are removed from the metadata,
and if only one sample group is found, it is renamed "All".}

\item{temp_dir}{The path to a directory where peak files will be 
downloaded.}

\item{force}{A logical indicating whether already present files be 
redownloaded.}
}
\value{
An object of class \linkS4class{ENCODEBindingConsensus}.
}
\description{
This function takes the result of a previous call to 
\code{\link{queryEncode}}, splits the contained peak files by conditions (as
specified by the \code{split_by} argument), then builds consensus peaks for 
each condition.
}
\examples{
query_results = queryEncodeGeneric(biosample_name="A549", assembly="GRCh38",
                                  file_format="^bed$", output_type="^peaks$", 
                                  treatment_duration_unit="minute",
                                  treatment_duration="(^5$|^10$)", 
                                  target="NR3C1", fixed=FALSE)
res = buildConsensusPeaks(query_results, split_by=c("treatment_duration"), 
                          consensus_threshold=0.5)

}
