% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/summaryBuilders.R
\name{buildExpressionMean}
\alias{buildExpressionMean}
\title{Calculates average expression levels of the results of a previously
completed ENCODE query.}
\usage{
buildExpressionMean(query_results, split_by, metric = NULL,
  temp_dir = ".", force = FALSE)
}
\arguments{
\item{query_results}{A data.table returned by \code{\link{queryEncode}} or 
\code{\link{queryEncodeGeneric}}.}

\item{split_by}{A vector of column names from query_results that will be used
to split the average expression levels.
If \code{NULL}, all elements of query_results are used in the
same average expression calculation.}

\item{temp_dir}{The path to a directory where peak files will be 
downloaded.}

\item{force}{A logical indicating whether already present files be 
redownloaded.}
}
\value{
An object of class \linkS4class{ENCODEExpressionSummary}.
}
\description{
This function takes the result of a previous call to 
\code{\link{queryEncode}}, splits the contained expression files by 
conditions (as specified by the \code{split_by} argument), then calculates
average expression levels for each condition.
}
