% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/rest_api.R
\name{clean_table}
\alias{clean_table}
\title{Clean a data.frame that was produced by ENCODE REST API}
\usage{
clean_table(table)
}
\arguments{
\item{table}{The table produced by the \code{extract_table} function.}
}
\value{
a \code{data.frame} corresponding to the cleaned version of the
input \code{data.frame}.
}
\description{
\code{data.frame}s produced when converting JSON to \code{data.frame} with
the \code{fromJSON} function will sometime have columns that are lists
and/or columns that are \code{data.frames}.
}
\details{
This function will either remove columns that are not relevant and convert
columns to a vector or data.frame.
}
