% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/GenomicOverlaps.R
\name{combined_regions,GenomicOverlaps-method}
\alias{combined_regions,GenomicOverlaps-method}
\title{Returns the combined regions from a \linkS4class{GenomicOverlaps} object.}
\usage{
\S4method{combined_regions}{GenomicOverlaps}(x)
}
\arguments{
\item{x}{The \linkS4class{GenomicOverlaps} object.}
}
\value{
A \code{GRanges} object representing the combined regions.
}
\description{
Returns the combined regions from a \linkS4class{GenomicOverlaps} object.
}
\keyword{internal}
