% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ENCODESummary.R
\docType{methods}
\name{consensus}
\alias{consensus}
\alias{consensus,ENCODEBindingConsensus-method}
\alias{consensus,ENCODEBindingConsensus,ENCODEBindingConsensus-method}
\title{Returns a \linkS4class{GRangesList} of the per-condition consensus peaks.}
\usage{
consensus(x)

\S4method{consensus}{ENCODEBindingConsensus}(x)
}
\arguments{
\item{x}{The \linkS4class{ENCODEBindingConsensus} object.}
}
\value{
A \linkS4class{GRangesList} of the per-condition consensus peaks.
}
\description{
Returns a \linkS4class{GRangesList} of the per-condition consensus peaks.
}
