% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/GenomicOverlaps.R
\name{consensus_indices}
\alias{consensus_indices}
\title{Determines which regions form a "consensus" from all input regions.}
\usage{
consensus_indices(x, consensus_threshold)
}
\arguments{
\item{x}{A \linkS4class{GenomicOverlaps} object.}

\item{consensus_threshold}{The fraction of input regions which must have
a given region for it to be selected.}
}
\value{
A vector of the numeric indices of the element of \code{regions(x)} 
        which fit the given criteria.
}
\description{
Determines which regions form a "consensus" from all input regions.
}
\keyword{internal}
