% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/createDesign.R
\name{createDesign}
\alias{createDesign}
\title{Create a design for the files associated with the result of a queryEncode,
searchEncode or fuzzySearch research.}
\usage{
createDesign(input = NULL, df = NULL, split = FALSE, fileFormat = "bam",
  dataset_type = "experiments", format = "long",
  output_type = "data.table", ID = c(1, 2))
}
\arguments{
\item{input}{The \code{data.frame} created by a queryEncode or
searchEncode research.}

\item{df}{The \code{data.frame} used to extract the files link.
Default :\code{NULL}}

\item{split}{Allow to the function to return a \code{list} of \code{data.frame}
where each \code{data.frame} contain the files for a single experiment
Default: \code{FALSE}.}

\item{fileFormat}{A string that correspond to the type of the files 
that need to be extract.
Default: bam}

\item{dataset_type}{A string that correspong to the type of dataset that
will be extrated.
Default: experiments}

\item{format}{The format (long or wide) to represent the data. The 'long' format
will contain three columns (File, Experiment, Value). The 'wide' format
organize the data as an array with the experiments as columns and files as rows.
Default: long}

\item{output_type}{The type of output of the function, can be \code{data.frame}
or a \code{data.table}
Default: \code{data.frame}}

\item{ID}{A two element numeric vector, that first element is the value assign
to replicate and the second is the value assign to control.
Default: 1 and 2}
}
\value{
is a \code{data.frame} with files for all the experiments or 
a \code{list} of \code{data.frame} with all the file per experiment when
the parameter split is set to \code{TRUE}
}
\description{
Create a design for the files associated with the result of a queryEncode,
searchEncode or fuzzySearch research.
}

