% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/download.R
\name{download_dt_file}
\alias{download_dt_file}
\title{Downloads all files inside a data.table.}
\usage{
download_dt_file(input_dt, dir, force, show_experiment = FALSE)
}
\arguments{
\item{input_dt}{The data.table to be looped over for determining which
files should be downloaded.}

\item{dir}{The path of the directory where the downloaded files should 
be saved.}

\item{force}{If TRUE, existing files are downloaded again.}

\item{show_experiment}{If TRUE, the name of the experiment is extracted
from the data table and displayed in status messages.}
}
\value{
The name of the files which were downloaded.
}
\description{
Downloads all files inside a data.table.
}
