% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/download.R
\name{download_single_file}
\alias{download_single_file}
\title{download_single_file Downloads a single file and checks if md5 checksums match.}
\usage{
download_single_file(file_url, file_md5, dir = ".",
  experiment_name = NULL, force = TRUE)
}
\arguments{
\item{file_url}{A \code{character} giving the URL of the file to be downloaded.}

\item{file_md5}{A \code{character} giving the expected md5 checksum hash of the
file to be downloaded.}

\item{dir}{The directory where the downloaded file should be saved. Default: "."}

\item{experiment_name}{An optional experiment name to be displayed with the 
status reports.}

\item{force}{\code{boolean} indicating if existing files should be downloaded 
again. Default : TRUE}
}
\value{
A \code{character} with the name of the downloaded file.
}
\description{
download_single_file Downloads a single file and checks if md5 checksums match.
}
