% Generated by roxygen2 (4.1.0): do not edit by hand
% Please edit documentation in R/ENCODEdb.R
\docType{data}
\name{encode_df}
\alias{encode_df}
\title{Metadata from the ENCODE database in a R object.
Super list de la mort qui tue !!}
\format{A list of data frames}
\usage{
data(encode_df)
}
\description{
Metadata from the ENCODE database in a R object.
Super list de la mort qui tue !!
}
\keyword{datasets}

