% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ENCODExplorer.R
\docType{data}
\name{encode_df_ext_2}
\alias{encode_df_ext_2}
\title{Extended metadata (2) from the ENCODE database in a R object.
\code{get_encode_df_full} should be used instead of accessing 
this object directly.}
\format{A data table}
\usage{
data(encode_df_ext_2)
}
\value{
A data table
}
\description{
Extended metadata (2) from the ENCODE database in a R object.
\code{get_encode_df_full} should be used instead of accessing 
this object directly.
}
\seealso{
\code{\link{get_encode_df_full}}, \code{\link{export_ENCODEdb_matrix_lite}}.
}
\keyword{datasets}
