\name{ENCODEdb-package}
\alias{ENCODEdb-package}
\alias{ENCODEdb}
\docType{package}
\title{
A compilation of ENCODE metadata.
}
\description{
This package allows user to quickly access ENCODE project files metadata and
give access to helper functions to query the ENCODE rest api.
}
\details{
\tabular{ll}{
Package: \tab ENCODEdb\cr
Type: \tab Package\cr
Version: \tab 0.99.0\cr
Date: \tab 2015-02-25\cr
License: \tab Artistic-2.0\cr
}
~~ An overview of how to use the package, including the most important ~~
~~ functions ~~
}
\author{
Charles Joly Beauparlant and Audrey Lemacon.

Maintainer: Who to complain to <charles.joly-beauparlant@crchul.ulaval.ca>
}
\references{
www.encodeproject.org
https://www.encodeproject.org/help/rest-api/
}
ENCODE, rest, api, metadata
\keyword{ package }
\examples{
~~ simple examples of the most important functions ~~
}
