% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ENCODESummary.R
\docType{class}
\name{ENCODEExpressionSummary-class}
\alias{ENCODEExpressionSummary-class}
\title{ENCODEExpressionSummary objects represents means of gene- or 
transcript-levels of expression across a set of ENCODE files.}
\description{
ENCODEExpressionSummary objects represents means of gene- or 
transcript-levels of expression across a set of ENCODE files.
}
\section{Slots}{

\describe{
\item{\code{raw_data}}{A list of data-frames containing the full raw data of each
of the downloaded ENCODE files.}

\item{\code{metric}}{A character giving the regular expression used to extract 
expression metrics from the ENCODE files.}

\item{\code{metric_data}}{A \link{data.frame} of the per-condition metric values.}

\item{\code{expression_type}}{The type of expression which is being reported, either 
gene or transcripts.}
}}

