% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/prepare_data.R
\name{export_ENCODEdb_matrix}
\alias{export_ENCODEdb_matrix}
\title{Extract essential informations from a list of data.table in a \code{list} of 
\code{data.table}s}
\usage{
export_ENCODEdb_matrix(database_filename)
}
\arguments{
\item{database_filename}{The name of the file to save the database into.}
}
\value{
a \code{list} containing two elements. The first one 'experiment' is 
a \code{data.table} containing essential informations for each file part of    
an experiment ; the second one 'dataset' is a \code{data.table} containing 
essential informations for each file part of a dataset.
}
\description{
Extract essential informations from a list of data.table in a \code{list} of 
\code{data.table}s
}
\examples{
    database_filename <- system.file("extdata/tables",package = "ENCODExplorer")
    \dontrun{
        export_ENCODEdb_matrix(database_filename = database_filename)
    }
}
