% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/prepare_data.R
\name{export_ENCODEdb_matrix_lite}
\alias{export_ENCODEdb_matrix_lite}
\title{Extract file metadata from the full set of ENCODE metadata tables.}
\usage{
export_ENCODEdb_matrix_lite(database_filename)
}
\arguments{
\item{database_filename}{A list of ENCODE metadata tables as loaded by
prepare_ENCODEdb.}
}
\value{
a \code{data.table} containing relevant metadata for all
  ENCODE files.
}
\description{
Extract file metadata from the full set of ENCODE metadata tables.
}
\examples{
    \dontrun{
        tables = prepare_ENCODEdb()
        export_ENCODEdb_matrix_lite(database_filename = tables)
    }
}
