% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ENCODESummary.R
\docType{methods}
\name{file_metadata,ENCODESummary-method}
\alias{file_metadata,ENCODESummary-method}
\title{Returns a list of per-condition metadata of the ENCODE files used to build
the \linkS4class{ENCODESummary} object.}
\usage{
\S4method{file_metadata}{ENCODESummary}(x)
}
\arguments{
\item{x}{The \linkS4class{ENCODESummary} object.}
}
\value{
A \code{list} of per-condition metadata of the ENCODE files used to 
build the \linkS4class{ENCODESummary} object.
}
\description{
Returns a list of per-condition metadata of the ENCODE files used to build
the \linkS4class{ENCODESummary} object.
}
