% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ENCODESummary.R
\docType{methods}
\name{fpkm,ENCODEExpressionSummary-method}
\alias{fpkm,ENCODEExpressionSummary-method}
\title{Returns a \link{data.frame} of the per-condition 
fragments per kilobase of exon model per million reads mapped (FPKM).}
\usage{
\S4method{fpkm}{ENCODEExpressionSummary}(x)
}
\arguments{
\item{x}{The \linkS4class{ENCODEExpressionSummary} object.}
}
\value{
A \link{data.frame} of the per-condition fragments per kilobase
        of exon model per million reads mapped (FPKM).
}
\description{
Returns a \link{data.frame} of the per-condition 
fragments per kilobase of exon model per million reads mapped (FPKM).
}
