% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/fuzzySearch.R
\name{fuzzySearch}
\alias{fuzzySearch}
\title{Fuzzysearch is a searching function for a string or a list of string 
within the encode_df \code{data.table}}
\usage{
fuzzySearch(searchTerm = NULL, database = NULL, filterVector = NULL,
  multipleTerm = FALSE, ignore_case = TRUE, fixed = FALSE)
}
\arguments{
\item{searchTerm}{The keyword or a \code{list} of keyword to search.}

\item{database}{A \code{data.table}}

\item{filterVector}{}

\item{multipleTerm}{A boolean that indicate if the searchTerm is a list or
even multiple searchTerm separete by a comma in a single string.}

\item{fixed}{Boolean use to apply a strict search for the searchTerm within
the database}
}
\value{
A \code{data.table} corresponding the every row of the database that
contain at least of one the searchTerm.
}
\description{
Fuzzysearch is a searching function for a string or a list of string 
within the encode_df \code{data.table}
}

