% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/GenomicOverlaps.R
\docType{class}
\name{GenomicOverlaps-class}
\alias{GenomicOverlaps-class}
\title{GenomicOverlaps objects represent the overlaps of multiple GRanges.}
\description{
GenomicOverlaps objects represent the overlaps of multiple GRanges.
}
\section{Slots}{

\describe{
\item{\code{regions}}{A GRanges object representing the combined regions.}

\item{\code{matrix}}{A matrix representing which elements overlap with the combined 
regions.}
}}

\keyword{internal}
