% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/GenomicOverlaps.R
\name{GenomicOverlaps}
\alias{GenomicOverlaps}
\title{Create a \linkS4class{GenomicOverlaps} object.}
\usage{
GenomicOverlaps(grl, import_spec = list())
}
\arguments{
\item{grl}{The \linkS4class{GRangesList} object whose elements need to be 
overlapped with each others.}

\item{import_spec}{A list of columns which should be imported into
the resulting object. Each element should be named after
a column in \code{mcols(grl)}, and should contain a 
function to be used to aggregate multiple values of that 
column.}
}
\value{
An object of class \code{GenomicOverlaps}.
}
\description{
Given a \linkS4class{GRangesList} object, determines which regions overlap 
with each others. The input regions are "flattened", and all overlapping
regions (within and across the elements of the input \code{grl} parameter) 
are mapped to a new region whose \code{start} and \code{end} are the minimum 
\code{start} and maximum \code{end} of the initial overlapping regions.
}
\details{
A matrix indicating which input regions fall within these new mapped regions 
is then produced.
}
\keyword{internal}
