% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/prepare_data.R
\name{get_encode_df}
\alias{get_encode_df}
\title{Returns a "light" version of ENCODE file metadata.}
\usage{
get_encode_df()
}
\value{
a \code{data.table} containing the most relevant
  metadata for all ENCODE files.
}
\description{
Returns a "light" version of ENCODE file metadata.
}
\examples{
    my_encode_df = get_encode_df()

}
