% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/demo.R
\name{get_encode_df_demo}
\alias{get_encode_df_demo}
\title{Get a demo encode_df}
\usage{
get_encode_df_demo()
}
\value{
A vector of regions filenames
}
\description{
Get a demo encode_df
}
\examples{

encode_df_demo <- get_encode_df_demo()

}
