% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/prepare_data.R
\name{get_encode_df_full}
\alias{get_encode_df_full}
\title{Concatenates all available file metadata into a single data table.}
\usage{
get_encode_df_full()
}
\value{
a \code{data.table} containing relevant metadata for all
  ENCODE files.
}
\description{
Concatenates all available file metadata into a single data table.
}
\examples{
    my_full_encode_df = get_encode_df_full()

}
