% Generated by roxygen2 (4.1.0): do not edit by hand
% Please edit documentation in R/prepare_data.R
\name{get_encode_types}
\alias{get_encode_types}
\title{A list of known tables from ENCODE database.}
\usage{
get_encode_types()
}
\value{
a vector of \code{character} with the names of the known tables in
  the ENCODE database.
}
\description{
The list was extracted from:
  https://github.com/ENCODE-DCC/encoded/tree/master/src/encoded/schemas
}
\examples{
types <- get_encode_types()
}

