% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/rest_api.R
\name{get_schemas}
\alias{get_schemas}
\title{Extract the schemas from ENCODE's github}
\usage{
get_schemas()
}
\value{
a \code{list} of schemas.
}
\description{
The JSONs are fetched from:
        https://github.com/ENCODE-DCC/encoded/tree/master/src/encoded/schemas
}
\details{
The data is extracted using the github api:
        https://developer.github.com/guides/getting-started/

The data is then downloaded using the \code{jsonlite} package.
}

