% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/GenomicOverlaps.R
\name{initial_regions,GenomicOverlaps-method}
\alias{initial_regions,GenomicOverlaps-method}
\title{Returns the initial regions from a \linkS4class{GenomicOverlaps} object.}
\usage{
\S4method{initial_regions}{GenomicOverlaps}(x)
}
\arguments{
\item{x}{The \linkS4class{GenomicOverlaps} object.}
}
\value{
A \code{GRangesList} object representing the initial regions.
}
\description{
Returns the initial regions from a \linkS4class{GenomicOverlaps} object.
}
\keyword{internal}
