% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/GenomicOverlaps.R
\name{intersect_matrix}
\alias{intersect_matrix}
\title{Returns the intersection matrix from a \linkS4class{GenomicOverlaps} object.}
\usage{
intersect_matrix(x)
}
\arguments{
\item{x}{The \linkS4class{GenomicOverlaps} object.}
}
\value{
A matrix with has as many columns as the number of items in the 
        initial \linkS4class{GRangesList}, and as many rows as the number of 
        combined regions. Its values indicate how many regions from the 
        initial \linkS4class{GRangesList} element map to each individual
        combined range.
}
\description{
Returns the intersection matrix from a \linkS4class{GenomicOverlaps} object.
}
\keyword{internal}
