% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/consensus_binding.R
\docType{methods}
\name{length,ENCODEBindingConsensus-method}
\alias{length,ENCODEBindingConsensus-method}
\title{Returns the number of elements of a \linkS4class{ENCODEBindingConsensus}
object.}
\usage{
\S4method{length}{ENCODEBindingConsensus}(x)
}
\arguments{
\item{x}{The \linkS4class{ENCODEBindingConsensus} object.}
}
\value{
The number of elements in \code{x}.
}
\description{
Returns the number of elements of a \linkS4class{ENCODEBindingConsensus}
object.
}
