% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ENCODESummary.R
\docType{methods}
\name{length,ENCODESummary-method}
\alias{length,ENCODESummary-method}
\title{Returns the number of elements of a \linkS4class{ENCODESummary}
object.}
\usage{
\S4method{length}{ENCODESummary}(x)
}
\arguments{
\item{x}{The \linkS4class{ENCODESummary} object.}
}
\value{
The number of elements in \code{x}.
}
\description{
Returns the number of elements of a \linkS4class{ENCODESummary}
object.
}
