% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/GenomicOverlaps.R
\docType{methods}
\name{length,GenomicOverlaps-method}
\alias{length,GenomicOverlaps-method}
\title{Returns the number of elements of a \linkS4class{GenomicOverlaps} object.}
\usage{
\S4method{length}{GenomicOverlaps}(x)
}
\arguments{
\item{x}{The \linkS4class{GenomicOverlaps} object.}
}
\value{
The number of elements in \code{x}.
}
\description{
Returns the number of elements of a \linkS4class{GenomicOverlaps} object.
}
\keyword{internal}
