% Generated by roxygen2 (4.1.0): do not edit by hand
% Please edit documentation in R/ENCODEdb.R
\docType{data}
\name{matrices}
\alias{matrices}
\title{Metadata from the ENCODE database in a R object.}
\format{A list of data frames}
\usage{
data(matrices)
}
\description{
Metadata from the ENCODE database in a R object.
}
\keyword{datasets}

