% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/consensus_binding.R
\docType{methods}
\name{metadata,ENCODEBindingConsensus-method}
\alias{metadata,ENCODEBindingConsensus-method}
\title{Returns a list of per-condition metadata of the ENCODE files used to build
the \linkS4class{ENCODEBindingConsensus} object.}
\usage{
\S4method{metadata}{ENCODEBindingConsensus}(x)
}
\arguments{
\item{x}{The \linkS4class{ENCODEBindingConsensus} object.}
}
\value{
A \code{list} of per-condition metadata of the ENCODE files used to 
build the \linkS4class{ENCODEBindingConsensus} object.
}
\description{
Returns a list of per-condition metadata of the ENCODE files used to build
the \linkS4class{ENCODEBindingConsensus} object.
}
