% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ENCODESummary.R
\docType{methods}
\name{metadata}
\alias{metadata}
\alias{metadata,ENCODESummary-method}
\alias{metadata,ENCODESummary,ENCODESummary-method}
\title{Returns a data-frame list of the common per-condition metadata of the ENCODE 
files used to build the \linkS4class{ENCODESummary} object.}
\usage{
metadata(x)

\S4method{metadata}{ENCODESummary}(x)
}
\arguments{
\item{x}{The \linkS4class{ENCODESummary} object.}
}
\value{
A \code{data.frame} of the common per-condition metadata of the 
ENCODE files used to build the \linkS4class{ENCODESummary} object.
}
\description{
Returns a data-frame list of the common per-condition metadata of the ENCODE 
files used to build the \linkS4class{ENCODESummary} object.
}
