% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ENCODESummary.R
\docType{methods}
\name{metric}
\alias{metric}
\alias{metric,ENCODEExpressionSummary-method}
\alias{metric,ENCODEBindingConsensus,ENCODEBindingConsensus-method}
\title{Returns the regular expression used to extract metric values from the 
ENCODE files.}
\usage{
metric(x)

\S4method{metric}{ENCODEExpressionSummary}(x)
}
\arguments{
\item{x}{The \linkS4class{ENCODEExpressionSummary} object.}
}
\value{
The regular expression used to extract metric values from the 
ENCODE files.
}
\description{
Returns the regular expression used to extract metric values from the 
ENCODE files.
}
