% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ENCODESummary.R
\docType{methods}
\name{metric_data}
\alias{metric_data}
\alias{metric_data,ENCODEExpressionSummary-method}
\alias{metric_data,ENCODEBindingConsensus,ENCODEBindingConsensus-method}
\title{Returns a \link{data.frame} of the per-condition metric values.}
\usage{
metric_data(x)

\S4method{metric_data}{ENCODEExpressionSummary}(x)
}
\arguments{
\item{x}{The \linkS4class{ENCODEExpressionSummary} object.}
}
\value{
A \link{data.frame} of the per-condition metric values.
}
\description{
Returns a \link{data.frame} of the per-condition metric values.
}
