% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/consensus_binding.R
\docType{methods}
\name{names,ENCODEBindingConsensus-method}
\alias{names,ENCODEBindingConsensus-method}
\title{Returns the names of the elements of a \linkS4class{ENCODEBindingConsensus} 
object.}
\usage{
\S4method{names}{ENCODEBindingConsensus}(x)
}
\arguments{
\item{x}{The \linkS4class{ENCODEBindingConsensus} object.}
}
\value{
The names of the elements in \code{x}.
}
\description{
Returns the names of the elements of a \linkS4class{ENCODEBindingConsensus} 
object.
}
