% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ENCODESummary.R
\docType{methods}
\name{names,ENCODESummary-method}
\alias{names,ENCODESummary-method}
\title{Returns the names of the elements of a \linkS4class{ENCODESummary} 
object.}
\usage{
\S4method{names}{ENCODESummary}(x)
}
\arguments{
\item{x}{The \linkS4class{ENCODESummary} object.}
}
\value{
The names of the elements in \code{x}.
}
\description{
Returns the names of the elements of a \linkS4class{ENCODESummary} 
object.
}
