% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/GenomicOverlaps.R
\docType{methods}
\name{names,GenomicOverlaps-method}
\alias{names,GenomicOverlaps-method}
\title{Returns the names of the elements of a \linkS4class{GenomicOverlaps} object.}
\usage{
\S4method{names}{GenomicOverlaps}(x)
}
\arguments{
\item{x}{The \linkS4class{GenomicOverlaps} object.}
}
\value{
The names of the elements in \code{x}.
}
\description{
Returns the names of the elements of a \linkS4class{GenomicOverlaps} object.
}
\keyword{internal}
