% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ENCODESummary.R
\docType{methods}
\name{names<-,ENCODEBindingConsensus,character-method}
\alias{names<-,ENCODEBindingConsensus,character-method}
\title{Set names of the elements of a \linkS4class{ENCODEBindingConsensus} object.}
\usage{
\S4method{names}{ENCODEBindingConsensus,character}(x) <- value
}
\arguments{
\item{x}{The \linkS4class{ENCODEBindingConsensus} object.}

\item{value}{The new names for the elements of the 
\linkS4class{ENCODEBindingConsensus} object.}
}
\description{
Set names of the elements of a \linkS4class{ENCODEBindingConsensus} object.
}
