% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ENCODESummary.R
\docType{methods}
\name{names<-,ENCODEExpressionSummary,character-method}
\alias{names<-,ENCODEExpressionSummary,character-method}
\title{Set names of the elements of a \linkS4class{ENCODEExpressionSummary} object.}
\usage{
\S4method{names}{ENCODEExpressionSummary,character}(x) <- value
}
\arguments{
\item{x}{The \linkS4class{ENCODEExpressionSummary} object.}

\item{value}{The new names for the elements of the 
\linkS4class{ENCODEExpressionSummary} object.}
}
\value{
The renamed object.
}
\description{
Set names of the elements of a \linkS4class{ENCODEExpressionSummary} object.
}
