% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/GenomicOverlaps.R
\docType{methods}
\name{names<-,GenomicOverlaps,character-method}
\alias{names<-,GenomicOverlaps,character-method}
\title{Set names of the elements of a \linkS4class{GenomicOverlaps} object.}
\usage{
\S4method{names}{GenomicOverlaps,character}(x) <- value
}
\arguments{
\item{x}{The \linkS4class{GenomicOverlaps} object.}

\item{value}{The new names for the elements of the 
\linkS4class{GenomicOverlaps} object.}
}
\value{
A copy of the object.
}
\description{
Set names of the elements of a \linkS4class{GenomicOverlaps} object.
}
\keyword{internal}
