% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/GenomicOverlaps.R
\name{pairwise_overlap}
\alias{pairwise_overlap}
\title{Calculate the pairwise overlaps of all factors within a 
\linkS4class{GenomicOverlaps} object.}
\usage{
pairwise_overlap(x)
}
\arguments{
\item{x}{A \linkS4class{GenomicOverlaps} object.}
}
\value{
A matrix containing the pairwise overlaps of all elements in
\code{x}. The row's element is used as the denominator. Therefore, the
matrix is not symmetric.
}
\description{
Calculate the pairwise overlaps of all factors within a 
\linkS4class{GenomicOverlaps} object.
}
\keyword{internal}
