% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ENCODESummary.R
\docType{methods}
\name{peaks}
\alias{peaks}
\alias{peaks,ENCODEBindingConsensus-method}
\alias{peaks,ENCODEBindingConsensus,ENCODEBindingConsensus-method}
\title{Returns a \code{list} of \linkS4class{GRangesList} of the per-condition 
original peaks used to build the \linkS4class{ENCODEBindingConsensus} object.}
\usage{
peaks(x)

\S4method{peaks}{ENCODEBindingConsensus}(x)
}
\arguments{
\item{x}{The \linkS4class{ENCODEBindingConsensus} object.}
}
\value{
A \code{list} of \linkS4class{GRangesList} of the per-condition  
        original peaks usedto build the \linkS4class{ENCODEBindingConsensus} 
        object.
}
\description{
Returns a \code{list} of \linkS4class{GRangesList} of the per-condition 
original peaks used to build the \linkS4class{ENCODEBindingConsensus} object.
}
