% Generated by roxygen2 (4.1.0): do not edit by hand
% Please edit documentation in R/prepare_data.R
\name{prepare_data}
\alias{prepare_data}
\title{Create the data.frames for the table in ENCODE}
\usage{
prepare_data()
}
\value{
a \code{list} with selected tables from ENCODE.

The following table are downloaded"
\describe{
  \item{biosample}{}
  \item{experiment}{}
  \item{lab}{}
  \item{dataset}{}
  \item{files}{}
  \item{target}{}
}
}
\description{
Create the data.frames for the table in ENCODE
}
\examples{
lab <- ENCODEdb:::extract_table("lab")
}

