% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/prepare_data.R
\name{prepare_ENCODEdb}
\alias{prepare_ENCODEdb}
\title{Create the list of data.table for the tables in ENCODE}
\usage{
prepare_ENCODEdb(database_filename = "tables.RDA",
  types = get_encode_types(), overwrite = FALSE)
}
\arguments{
\item{database_filename}{The name of the file to save the database into.}

\item{types}{The names of the tables to extract using the ENCODE rest api.}

\item{overwrite}{If database_filename already exists, should it be overwritten?
Default: \code{FALSE}.}
}
\value{
is a \code{list} with selected tables from ENCODE.
}
\description{
Create the list of data.table for the tables in ENCODE
}
\examples{
prepare_ENCODEdb(database_filename = "tables.RDA", types = "platform")
file.remove("platform.RDA")
    \dontrun{
        prepare_ENCODEdb("ENCODEdb.RDA")
    }
    
}
