% Generated by roxygen2 (4.1.0): do not edit by hand
% Please edit documentation in R/query.R
\name{query}
\alias{query}
\title{Produce a subset of data following predefined criteria}
\usage{
query(df = NULL, accession = NULL, assay = NULL, biosample = NULL,
  dataset_access = NULL, file_accession = NULL, file_format = NULL,
  lab = NULL, organism = NULL, target = NULL, treatment = NULL,
  fixed = TRUE)
}
\arguments{
\item{df}{\code{list} of two \code{data.frame} containing ENCODE
experiment and dataset metadata}

\item{accession}{character string to select the experiment or dataset accession}

\item{assay}{character string to select the assay type}

\item{biosample}{character string to select the biosample name}

\item{dataset_access}{character string to select the dataset accession}

\item{file_accession}{character string to select the file accesion}

\item{file_format}{character string to select the file format}

\item{lab}{character string to select the laboratory}

\item{organism}{character string to select the donor organism}

\item{target}{character string to select the experimental target}

\item{treatment}{character string to select the treatment}

\item{fixed}{logical. If TRUE, pattern is a string to be matched as it is.}
}
\value{
a \code{list} of two \code{data.frame}s containing data about ENCODE
experiments and datasets
}
\description{
After running the \code{prepare_ENCODEDb} function, this function will allow
you to extract a subset of data encording to the following criteria :
accession, assay name, biosample, dataset accession, file accession,
file format, laboratory, donor organism, target and treatment.
}
\details{
By default, the query can be made on an exact match term. This behaviour can
be modified by setting the \code{fixed} argument at \code{TRUE}
}
\examples{
query(biosample = "A549", file_format = "bam")
}

