% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/consensus_binding.R
\name{queryConsensusPeaks}
\alias{queryConsensusPeaks}
\title{Queries ENCODE for consensus peaks.}
\usage{
queryConsensusPeaks(biosample_name, assembly, target)
}
\arguments{
\item{biosample_name}{The cell-line/tissue for which consensus peaks should 
be queried.}

\item{assembly}{The target genomic assembly.}

\item{target}{The target protein.}
}
\value{
An object of class \linkS4class{ENCODEBindingConsensus}.
}
\description{
Queries the ENCODE metadata to determine which peak files exists for the 
\code{target} protein in the \code{biosample_name} biosample for the 
\code{assembly} genomic assembly, then builds per-condition (as determined 
by the \code{treatment} column and its adjuncts) consensus peaks.
}
\details{
If you wish to have more control over the files used to build the consensus,
use \code{\link{buildConsensusPeaks}}.
}
\seealso{
\code{\link{buildConsensusPeaks}}
}
