% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/consensus_binding.R
\name{queryConsensusPeaksAll}
\alias{queryConsensusPeaksAll}
\title{Queries and returns consensus peaks for all available targets of a given 
biosample_name.}
\usage{
queryConsensusPeaksAll(biosample_name, assembly)
}
\arguments{
\item{biosample_name}{The cell-line/tissue for which consensus peaks should 
be queried.}

\item{assembly}{The target genomic assembly.}
}
\value{
A list of \linkS4class{ENCODEBindingConsensus}, one per available 
        target.
}
\description{
This utility function calls \link{queryConsensusPeaks} in a loop over all
available targets for a given biosample_name, and returns the results in a 
list.
}
\seealso{
\code{\link{buildConsensusPeaks}}, \code{\link{queryConsensusPeaks}}
}
