% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ENCODESummary.R
\docType{methods}
\name{raw_data}
\alias{raw_data}
\alias{raw_data,ENCODEExpressionSummary-method}
\alias{raw_data,ENCODEBindingConsensus,ENCODEBindingConsensus-method}
\title{Returns the regular expression used to extract raw_data values from the 
ENCODE files.}
\usage{
raw_data(x)

\S4method{raw_data}{ENCODEExpressionSummary}(x)
}
\arguments{
\item{x}{The \linkS4class{ENCODEExpressionSummary} object.}
}
\value{
A list with one element per file-group containing a list of the 
        raw imported ENCODE files.
}
\description{
Returns the regular expression used to extract raw_data values from the 
ENCODE files.
}
