% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/query.R
\name{resolveEncodeAccession}
\alias{resolveEncodeAccession}
\title{Return a \code{data.frame} containing basic datasets information from an 
accession number}
\usage{
resolveEncodeAccession(accession)
}
\arguments{
\item{accession}{character, dataset accession number}
}
\value{
a \code{data.frame} containing basic datasets information for the 
requested accession number
}
\description{
Return a \code{data.frame} containing basic datasets information from an 
accession number
}
\examples{
res <- resolveEncodeAccession(accession = 'ENCSR361ONJ')$accession

}

