% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/rest_api.R
\name{searchEncode}
\alias{searchEncode}
\title{Simulate a query on ENCODE website and return the result as a 
\code{data.frame}}
\usage{
searchEncode(searchTerm = NULL, limit = 10, username = NULL,
  password = NULL)
}
\arguments{
\item{searchTerm}{a search term}

\item{limit}{the maximum number of return entries, default 10. \code{limit = "all"}
will return all the result. It can generate large results set.}

\item{username}{username}

\item{password}{password}
}
\value{
a \code{data.frame} corresponding every object that matches the 
search term
}
\description{
\code{data.frame}s produced when converting JSON to \code{data.frame} with
the \code{fromJSON} function will sometime have columns that are lists
and/or columns that are \code{data.frames}.
}
\details{
This function simulates a basic query on ENCODE website
}
\examples{
 searchEncode("ChIP-Seq+H3K4me1")
}

