% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/query.R
\name{searchToquery}
\alias{searchToquery}
\title{Convert searchEncode output in queryEncode output.}
\usage{
searchToquery(df = NULL, searchResults, quiet = TRUE)
}
\arguments{
\item{df}{\code{list} of two \code{data.frame} containing ENCODE 
experiment and dataset metadata.}

\item{searchResults}{the results set generated from \code{searchEncode}}

\item{quiet}{logical enables to switch off the result summary information 
when setting at TRUE.}
}
\value{
a \code{list} of two \code{data.frame}s containing data about ENCODE 
experiments and datasets
}
\description{
After processing to a basic search with the \code{searchEncode} function 
you can convert your result in a queryEncode output. Thus you can benefit 
from the collected metadata.
}
\details{
The output is compatible with the dowload function.
}
\examples{
search_res <- searchEncode(searchTerm = "switchgear elavl1", limit = "1")
res <- searchToquery(searchResults = search_res, quiet = TRUE)

}
