% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/shiny_encode.R
\name{shinyEncode}
\alias{shinyEncode}
\title{Launch a shiny interface for ENCODExplorer}
\usage{
shinyEncode()
}
\value{
None
}
\description{
Launch a shiny interface for ENCODExplorer
}
\examples{
 \dontrun{shinyEncode}

}
