% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ENCODESummary.R
\docType{methods}
\name{show,ENCODEBindingConsensus-method}
\alias{show,ENCODEBindingConsensus-method}
\title{Prints a summary of a \linkS4class{ENCODEBindingConsensus} object.}
\usage{
\S4method{show}{ENCODEBindingConsensus}(object)
}
\arguments{
\item{object}{The \linkS4class{ENCODEBindingConsensus} object.}
}
\description{
Prints a summary of a \linkS4class{ENCODEBindingConsensus} object.
}
