% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ENCODESummary.R
\docType{methods}
\name{show,ENCODEExpressionSummary-method}
\alias{show,ENCODEExpressionSummary-method}
\title{Prints a summary of a \linkS4class{ENCODEExpressionSummary} object.}
\usage{
\S4method{show}{ENCODEExpressionSummary}(object)
}
\arguments{
\item{object}{The \linkS4class{ENCODEExpressionSummary} object.}
}
\value{
Nothing.
}
\description{
Prints a summary of a \linkS4class{ENCODEExpressionSummary} object.
}
