% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/summaryBuilders.R
\name{split_by_metadata}
\alias{split_by_metadata}
\title{Given a metadata data-frame and a vector of column names (split_by),
builds a partition of rows based on all possible combinations
of the columns identified by split_by.}
\usage{
split_by_metadata(metadata, split_by)
}
\arguments{
\item{metadata}{A data.frame with the metadata.}

\item{split_by}{The names of the columns in metadata according to which they 
should be split.}
}
\value{
A list with information about the new partition.
}
\description{
Given a metadata data-frame and a vector of column names (split_by),
builds a partition of rows based on all possible combinations
of the columns identified by split_by.
}
\keyword{data.table}
\keyword{importFrom}
\keyword{internal}
\keyword{rbindlist}
