% Generated by roxygen2 (4.1.0): do not edit by hand
% Please edit documentation in R/ENCODEdb.R
\docType{data}
\name{tables}
\alias{tables}
\title{List of selected tables from ENCODE rest api.}
\format{A list of data frames}
\usage{
data(tables)
}
\description{
/itemize{
  \item biosample
  \item experiment
  \item lab
  \item dataset
  \item files
  \item target
}
}
\keyword{datasets}

