% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ENCODESummary.R
\docType{methods}
\name{tpm,ENCODEExpressionSummary-method}
\alias{tpm,ENCODEExpressionSummary-method}
\title{Returns a \link{data.frame} of the per-condition transcript-per-millions 
(TPM).}
\usage{
\S4method{tpm}{ENCODEExpressionSummary}(x)
}
\arguments{
\item{x}{The \linkS4class{ENCODEExpressionSummary} object.}
}
\value{
A \link{data.frame} of the per-condition transcript-per-millions.
}
\description{
Returns a \link{data.frame} of the per-condition transcript-per-millions 
(TPM).
}
