% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/GenomicOverlaps.R
\name{union_indices}
\alias{union_indices}
\title{Calculate the union of a set of elements within a 
\linkS4class{GenomicOverlaps} object.}
\usage{
union_indices(x, indices = names(x))
}
\arguments{
\item{x}{A \linkS4class{GenomicOverlaps} object.}

\item{indices}{The names of the elements from \code{x} whose union 
should be found.}
}
\value{
A vector of the numeric indices of the element of \code{regions(x)} 
        which fit the given criteria.
}
\description{
Calculate the union of a set of elements within a 
\linkS4class{GenomicOverlaps} object.
}
\keyword{internal}
