% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/prepare_data.R
\name{update_ENCODExplorer}
\alias{update_ENCODExplorer}
\title{Create or update all the needed data for ENCODExplorer}
\usage{
update_ENCODExplorer(database_filename = "inst/extdata/ENCODEdb.sqlite",
  types = get_encode_types(), overwrite = FALSE)
}
\arguments{
\item{database_filename}{The name of the file to save the database into.}

\item{types}{The names of the tables to extract from ENCODE rest api.}

\item{overwrite}{Should tables already present in database be overwrited
Default: \code{FALSE}.}
}
\value{
none if \code{overwrite} is set to TRUE or return a \code{list} 
containg two objects encode_df and accession_df.
}
\description{
This function creates or updates ENCODExplorer data according the following 
steps :
1) Create the RSQLite databse for the tables in ENCODE
2) Extract essential informations from the RSQLite databse in encode_df
3) Extract accession numbers from all the datasets of RSQLite databse in 
accession_df
4) if overwrite = TRUE, the new encode_df will overwrite the former one else 
return the newly generated objets.
}
\examples{
update_ENCODExplorer(database_filename = "platform.sql", types = "platform")
file.remove("platform.sql")
  \dontrun{
    update_ENCODExplorer("ENCODEdb.sqlite")
  }
  
}

