% Generated by roxygen2 (4.1.0): do not edit by hand
% Please edit documentation in R/rest_api.R
\name{download}
\alias{download}
\title{Download files from the Internet.}
\usage{
download(resultSet, resultOrigin, format, dir)
}
\arguments{
\item{resultSet}{the results set}

\item{resultOrigin}{name of the function used to generate the result set
(\code{search} or \code{query})}

\item{format}{file format, default = all}

\item{dir}{the name of the directory where the downloaded file will be saved. Default = /tmp}
}
\description{
After processing to a basic search with the \code{search} function or a
precise search thanks to the \code{query} function, you can proceed to the
downloading of all the corresponding files.
}
\details{
This function can be used to download a set of files by providing the results
set, its origin (search or query), the file format and finally the destination
directory.
}
\examples{
query_results <- query(assay = "switchgear", target ="elavl1", file_format = "bed_broadPeak" , fixed = F)
download(resultSet = query_results, resultOrigin = "query")
}

