% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/download.R
\name{downloadEncode}
\alias{downloadEncode}
\title{downloadEncode is use to download  a serie of files or dataset 
by their accession.}
\usage{
downloadEncode(file_acc = NULL, df = NULL, format = "all", dir = ".",
  force = TRUE)
}
\arguments{
\item{file_acc}{A \code{character} of ENCODE file accession or 
experiment accession. Can also be a data.table coming from any ENCODExplorer
research function.}

\item{df}{The reference \code{data.table} use to find the download. File 
that are not available will be search directly throught the current
ENCODE database.}

\item{format}{The specific file format to download.
Default : all}

\item{dir}{The directory to locate the downloaded files}

\item{force}{\code{boolean} the allow to download a file when it already exist
in the directory. 
Default : TRUE}
}
\description{
downloadEncode is use to download  a serie of files or dataset 
by their accession.
}
\examples{
 \dontrun{downloadEncode(file_acc = "ENCFF567UCJ")}

}

