% Generated by roxygen2 (4.1.0): do not edit by hand
% Please edit documentation in R/download.R
\name{downloadEncode}
\alias{downloadEncode}
\title{Download files from the Internet.}
\usage{
downloadEncode(df = NULL, resultSet = NULL, resultOrigin = NULL,
  format = "all", dir = ".")
}
\arguments{
\item{df}{\code{list} of two \code{data.frame} containing ENCODE experiment
and dataset metadata.}

\item{resultSet}{the results set.}

\item{resultOrigin}{name of the function used to generate the result set
(\code{searchEncode} or \code{queryEncode})}

\item{format}{file format, default = all}

\item{dir}{the name of the directory where the downloaded file will be saved.
Default = current directory}
}
\value{
void
}
\description{
After processing to a basic search with the \code{searchEncode} function or a
precise search thanks to the \code{queryEncode} function, you can proceed to
the downloading of all the corresponding files.
}
\details{
This function can be used to download a set of files by providing the results
set, its origin (searchEncode or queryEncode), the file format and finally
the destination directory.
}
\examples{
resultSet <- queryEncode(biosample = "A549", file_format = "bam")
  \dontrun{
  downloadEncode(resultSet = resultSet, dir = ".")
  }
}

