% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ENCODESummary.R
\docType{class}
\name{ENCODEBindingConsensus-class}
\alias{ENCODEBindingConsensus-class}
\title{ENCODEBindingConsensus objects represents consensus peaks derived from a 
set of ENCODE files.}
\description{
ENCODEBindingConsensus objects represents consensus peaks derived from a 
set of ENCODE files.
}
\section{Slots}{

\describe{
\item{\code{peaks}}{The per-condition original peaks used to build the consensus.}

\item{\code{consensus}}{The per-condition consensus peaks.}

\item{\code{consensus_threshold}}{The proportion of replicates which must bear
a specific peak for it to be added to the set of
consensus peaks.}
}}

