% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/consensus_binding.R
\docType{class}
\name{ENCODEBindingConsensus-class}
\alias{ENCODEBindingConsensus-class}
\title{ENCODEBindingConsensus objects represents consensus peaks derived from a 
set of ENCODE files.}
\description{
ENCODEBindingConsensus objects represents consensus peaks derived from a 
set of ENCODE files.
}
\section{Slots}{

\describe{
\item{\code{metadata}}{A list of data-frames representing the ENCODE metadata
of the files used to build the per-condition consensus.}

\item{\code{peaks}}{The per-condition original peaks used to build the consensus.}

\item{\code{consensus}}{The per-condition consensus peaks.}
}}

