% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ENCODESummary.R
\docType{class}
\name{ENCODEExpressionSummary-class}
\alias{ENCODEExpressionSummary-class}
\title{ENCODEExpressionSummary objects represents means of gene- or 
transcript-levels of expression across a set of ENCODE files.}
\description{
ENCODEExpressionSummary objects represents means of gene- or 
transcript-levels of expression across a set of ENCODE files.
}
\section{Slots}{

\describe{
\item{\code{tpm}}{A \link{data.frame} of the per-condition transcript-per-millions.}

\item{\code{fpkm}}{A \link{data.frame} of the per-condition fragments per kilobase
of exon model per million reads mapped (FPKM).}
}}

