% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ENCODESummary.R
\docType{class}
\name{ENCODEExpressionSummary-class}
\alias{ENCODEExpressionSummary-class}
\alias{names<-,ENCODEExpressionSummary,character-method}
\alias{metric_data}
\alias{metric_data,ENCODEExpressionSummary-method}
\alias{metric_data,ENCODEBindingConsensus,ENCODEBindingConsensus-method}
\alias{metric}
\alias{metric,ENCODEExpressionSummary-method}
\alias{metric,ENCODEBindingConsensus,ENCODEBindingConsensus-method}
\alias{raw_data}
\alias{raw_data,ENCODEExpressionSummary-method}
\alias{raw_data,ENCODEBindingConsensus,ENCODEBindingConsensus-method}
\alias{show,ENCODEExpressionSummary-method}
\title{ENCODEExpressionSummary summarize means of expression across ENCODE files.}
\usage{
\S4method{names}{ENCODEExpressionSummary,character}(x) <- value

metric_data(x)

\S4method{metric_data}{ENCODEExpressionSummary}(x)

metric(x)

\S4method{metric}{ENCODEExpressionSummary}(x)

raw_data(x)

\S4method{raw_data}{ENCODEExpressionSummary}(x)

\S4method{show}{ENCODEExpressionSummary}(object)
}
\arguments{
\item{x}{The \linkS4class{ENCODESummary} object.}

\item{value}{The new names for the elements of the 
\linkS4class{ENCODESummary} object.}

\item{object}{The \linkS4class{ENCODESummary} object.}
}
\value{
For \code{raw_data}, a \code{list} of \linkS4class{GRangesList} 
  of the per-condition original expression tables used to build the object.
  For \code{metric}, the regular expression used to select the 
  column of metric values from the ENCODE files. For \code{metric_data},
  a \link{data.frame} of the per-condition metric values.
}
\description{
ENCODEExpressionSummary objects represent means (or medians) of expression
levels across multiple replicate samples, split by arbitrary metadata 
columns. They can be constructed using the \link{queryGeneExpression},
\link{queryTranscriptExpression} and \link{buildExpressionSummary} functions.
}
\section{Slots}{

\describe{
\item{\code{raw_data}}{A list of data-frames containing the full raw data of each
of the downloaded ENCODE files.}

\item{\code{metric}}{A character giving the regular expression used to extract 
expression metrics from the ENCODE files.}

\item{\code{metric_data}}{A \link{data.frame} of the per-condition metric values.}

\item{\code{expression_type}}{The type of expression which is being reported, either 
gene or transcripts.}
}}

\section{Methods}{


  ENCODEExpressionSummary object can be accessed through the methods from the 
  ENCODESummary class, as well as ENCODEBindingConsensus-specific methods:

  \describe{
     \item{\code{raw_data}}{Returns a \code{list} of \linkS4class{GRangesList} 
                            of the per-condition original expression tables
                            used to build the object.}
     \item{\code{metric}}{Returns the regular expression used to select the 
                          column of metric values from the ENCODE files.}
     \item{\code{metric_data}}{Returns a \link{data.frame} of the 
                               per-condition metric values.}
  }
}

\examples{
  res = queryGeneExpression("bone marrow")
  raw_data(res)
  metric(res)
  metric_data(res)

}
