% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ENCODESummary.R
\docType{class}
\name{ENCODESummary-class}
\alias{ENCODESummary-class}
\title{ENCODESummary objects represents a summary derived from multiple ENCODE 
files.}
\description{
ENCODESummary objects represents a summary derived from multiple ENCODE 
files.
}
\section{Slots}{

\describe{
\item{\code{files}}{The path of the files used in this summary.}

\item{\code{file_metadata}}{A list of data-frames representing the ENCODE metadata
of the files used to build the per-condition consensus.}

\item{\code{metadata}}{A data-frame with the metadata of each element in the summary.}
}}

