% Generated by roxygen2 (4.1.0): do not edit by hand
% Please edit documentation in R/prepare_data.R
\name{export_ENCODEdb_matrix}
\alias{export_ENCODEdb_matrix}
\title{Extract essential informations from the RSQLite databse in a \code{list} of
\code{data.frame}s}
\usage{
export_ENCODEdb_matrix(database_filename)
}
\arguments{
\item{database_filename}{The name of the file to save the database into.
Default: \code{\"ENCODEdb.sqlite\"}.}
}
\value{
a \code{list} containing two elements. The first one 'experiment' is
a \code{data.frame} containing essential informations for each file part of
an experiment ; the second one 'dataset' is a \code{data.frame} containing
essential informations for each file part of a dataset.
}
\description{
Extract essential informations from the RSQLite databse in a \code{list} of
\code{data.frame}s
}

