% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/prepare_data.R
\name{export_ENCODEdb_matrix}
\alias{export_ENCODEdb_matrix}
\title{Extract essential informations from the RSQLite databse in a \code{list} of 
\code{data.frame}s}
\usage{
export_ENCODEdb_matrix(database_filename)
}
\arguments{
\item{database_filename}{The name of the file to save the database into.
Default: \code{ENCODEdb.sqlite}.}
}
\value{
A \code{data.frame} containing essential informations for each file
part of an experiment.
}
\description{
Extract essential informations from the RSQLite databse in a \code{list} of 
\code{data.frame}s
}
\examples{
  database_filename <- system.file("extdata/ENCODEdb.sqlite",
                                   package = "ENCODEdb")
  \dontrun{
    export_ENCODEdb_matrix(database_filename = database_filename)
  }
}

