% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/rest_api.R
\name{extract_table}
\alias{extract_table}
\title{Extract a data.frame corresponding to a table in ENCODE database}
\usage{
extract_table(type)
}
\arguments{
\item{type}{The type of table to extract from ENCODE rest api.}
}
\value{
a \code{data.frame} corresponding to the table asked. If no match is
  found, returns an empty \code{data.frame}
}
\description{
Extract a data.frame corresponding to a table in ENCODE database
}

