% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/fuzzySearch.R
\name{fuzzySearch}
\alias{fuzzySearch}
\title{Fuzzysearch is a searching function for a string or a list of string 
within the encode_df \code{data.table}. For faster processing, pass encode_df
object as database parameter.}
\usage{
fuzzySearch(
  searchTerm = NULL,
  database = get_encode_df(),
  filterVector = NULL,
  multipleTerm = FALSE,
  ignore_case = TRUE
)
}
\arguments{
\item{searchTerm}{The keyword or a \code{list} of keyword to search.}

\item{database}{A \code{data.table} with similar format as encode_df database.}

\item{filterVector}{A \code{character} to apply the search on specific column.}

\item{multipleTerm}{A boolean that indicate if the searchTerm is a list or
even multiple searchTerm separete by a comma in a single string.}

\item{ignore_case}{A \code{boolean} to enable the case sensitivity.}
}
\value{
A \code{data.table} corresponding the every row of the database that
contain at least of one the searchTerm.
}
\description{
Fuzzysearch is a searching function for a string or a list of string 
within the encode_df \code{data.table}. For faster processing, pass encode_df
object as database parameter.
}
\examples{
 
   fuzz_ex <- fuzzySearch(searchTerm=c("ELAVL1","atf7"),
   database=get_encode_df_demo(), filterVector ="target", multipleTerm = TRUE)

}
