% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/prepare_data.R
\name{get_encode_types}
\alias{get_encode_types}
\title{A list of known tables from ENCODE database.}
\usage{
get_encode_types()
}
\value{
a vector of \code{character} with the names of the known tables in
  the ENCODE database.
}
\description{
The type (table) names are extracted from the schema list from ENCODE-DCC
github repository:
  https://github.com/ENCODE-DCC/encoded/tree/master/src/encoded/schemas
}
\details{
The data is extracted using the github api:
  https://developer.github.com/guides/getting-started/
}

