% Generated by roxygen2 (4.1.0): do not edit by hand
% Please edit documentation in R/schemas.R
\name{get_schemas}
\alias{get_schemas}
\title{Extract the schemas from ENCODE's github}
\usage{
get_schemas()
}
\value{
a \code{list} of JSON converted in \code{data.frame}.
}
\description{
The JSONs are fetched from:
  https://github.com/ENCODE-DCC/encoded/tree/master/src/encoded/schemas
}
\details{
The data is extracted using the github api:
  https://developer.github.com/guides/getting-started/

The data is then converted in \code{data.frame} using the \code{jsonlite}
package.
}
\examples{
schema <- get_schemas()
}

