% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/consensus_binding.R
\docType{methods}
\name{peaks,ENCODEBindingConsensus-method}
\alias{peaks,ENCODEBindingConsensus-method}
\title{Returns a \linkS4class{GRangesList} of the per-condition original peaks used 
to build the \linkS4class{ENCODEBindingConsensus} object.}
\usage{
\S4method{peaks}{ENCODEBindingConsensus}(x)
}
\arguments{
\item{x}{The \linkS4class{ENCODEBindingConsensus} object.}
}
\value{
A \linkS4class{GRangesList} of the per-condition original peaks used 
        to build the \linkS4class{ENCODEBindingConsensus} object.
}
\description{
Returns a \linkS4class{GRangesList} of the per-condition original peaks used 
to build the \linkS4class{ENCODEBindingConsensus} object.
}
