% Generated by roxygen2 (4.1.0): do not edit by hand
% Please edit documentation in R/prepare_data.R
\name{prepare_ENCODEdb}
\alias{prepare_ENCODEdb}
\title{Create the RSQLite databse for the tables in ENCODE}
\usage{
prepare_ENCODEdb(database_filename)
}
\arguments{
\item{database_filename}{The name of the file to save the database into.
Default: \code{\"ENCODEdb.sqlite\"}.}

\item{types}{The names of the tables to extract from ENCODE rest api.}

\item{overwrite}{Should tables already present in database be overwrited?
Default: \code{FALSE}.}
}
\value{
is a \code{list} with selected tables from ENCODE that were used to
create the \code{RSQLite} database.
}
\description{
Create the RSQLite databse for the tables in ENCODE
}

