% Generated by roxygen2 (4.1.0): do not edit by hand
% Please edit documentation in R/rest_api.R
\name{query}
\alias{query}
\title{Produce a subset of data following predefined criteria}
\usage{
query(accession = NULL, assay = NULL, biosample = NULL,
  dataset_access = NULL, file_accession = NULL, file_format = NULL,
  lab = NULL, organism = NULL, target = NULL, treatment = NULL,
  fixed = TRUE)
}
\arguments{
\item{selection}{criteria.}
}
\value{
a \code{list} of two \code{data.frame}s containing data about ENCODE
experiments and datasets
}
\description{
After running the \code{prepare_ENCODEDb} function, this function will allow
you to extract a subset of data encording to the following criteria :
accession, assay name, biosample, dataset accession, file accession,
file format, laboratory, donor organism, target and treatment.
}
\details{
By default, the query can be made on an exact match term. This behaviour can
be modified by setting the \code{fixed} argument at \code{TRUE}
}
\examples{
matrices = ENCODEdb:export_ENCODEDB_matrix(database_filename)
res = ENCODEdb::query(file_format = "GTF", biosample = "mcf-7", fixed = F)
}

