% Generated by roxygen2 (4.1.0): do not edit by hand
% Please edit documentation in R/ENCODEdb.R
\docType{data}
\name{tables}
\alias{tables}
\title{List of selected tables from ENCODE rest api.}
\format{A list of data frames}
\usage{
data(tables)
}
\description{
The tables were generated with the \code{prepare_ENCODEdb} function.
}
\details{
List of available tables:
\describe{
  \item{biosample}{}
  \item{experiment}{}
  \item{lab}{}
  \item{dataset}{}
  \item{files}{}
  \item{target}{}
  \item{organism}{}
  \item{replicate}{}
  \item{award}{}
  \item{treatment}{}
  \item{human_donor}{}
  \item{library}{}
  \item{platform}{}
  \item{document}{}
}
}
\keyword{datasets}

