% Generated by roxygen2 (4.1.0): do not edit by hand
% Please edit documentation in R/ENCODEdb.R
\docType{data}
\name{tables}
\alias{tables}
\title{List of selected tables from ENCODE rest api.}
\format{A list of data frames}
\usage{
data(tables)
}
\description{
The tables were generated with the \code{prepare_ENCODEdb} function.
}
\seealso{
\code{\link{get_encode_type}} to get a list of possible types. Note
  that some of the types are empty tables that are not included in the
  database created with \code{\link{prepare_ENCODEdb}} function.
}
\keyword{datasets}

